pyproject.toml
[project]
name = "quark-py"
version = "0.1.0"
description = "FastAPI + Ant Design Pro 前后端分离的管理后台，Build Anything With Quark"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "bcrypt>=4.3.0",
    "captcha>=0.7.1",
    "fastapi>=0.115.12",
    "fastapi-cache2>=0.2.2",
    "openpyxl>=3.1.5",
    "psutil>=7.0.0",
    "python-i18n>=0.3.9",
    "python-jose[cryptography]>=3.5.0",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple"
